// src/types/BirdCard.ts
export interface BirdCard {
  "Common name": string;
  "Scientific name": string;
  Set: string;
  Color: string;
  "Power text": string | null;
  Predator: boolean | null;
  Flocking: boolean | null;
  "Bonus card": boolean | null;
  "Victory points": number;
  "Nest type": string;
  "Egg limit": number;
  Wingspan: number;
  Forest: boolean | null;
  Grassland: boolean | null;
  Wetland: boolean | null;
  Invertebrate: number | null;
  Seed: number | null;
  Fish: number | null;
  Fruit: number | null;
  Rodent: number | null;
  Nectar: number | null;
  "Wild (food)": number | null;
  "/ (food cost)": boolean | null;
  "* (food cost)": boolean | null;
  "Total food cost": number;
  "Beak direction": string | null;
  id: number;
  "Native name": string | null;
  "Flavor text": string | null;
  "North America": boolean | null;
  "Central America": boolean | null;
  "South America": boolean | null;
  Europe: boolean | null;
  Asia: boolean | null;
  Africa: boolean | null;
  Oceania: boolean | null;
  Anatomist: boolean | null;
  Cartographer: boolean | null;
  Historian: boolean | null;
  Photographer: boolean | null;
}